Add files via upload
This is a modified version of nianeyna's original program, adapted by mkp (shamelesslymkp.tumblr.com). It provides some additional functionality - most notably, the ability to download marked for later items into an entirely separate downloads folder. The additional functionality required the creation of multiple additional scripts (well. the copying-and-adaptation of existing scripts, anyways. it's kind of like creation.) and some additions and modifications to the original scripts. Additional details about these changes have been added to the appropriate sections of this readme file and additions/changes to the original files have been marked with comments throughout.